+++ docs/docs/releases.md
@@ -13,6 +13,22 @@ language: 'en'
 - Fix: Underline cursor not working.
 - Fix:  sixel: Text doesn't overwrite sixels [#636](https://github.com/raphamorim/rio/issues/636).
 - Initial support to Sixel protocol.
+- **breaking:** Revamp the cursor configuration
+
+Before:
+
+```toml
+cursor = '▇'
+blinking-cursor = false
+```
+
+After:
+
+```toml
+[cursor]
+shape = 'block'
+blinking = false
+```
 
 ## 0.1.11